# docs/src/openapi.yml
swagger: 2.0
info:
  title: FHIR Server
  version: 0.0.1
  description: An FHIR server that is conformant to the Argonaut Data Query Implementation specification.
  termsOfService: https://www.google.com
  contact:
    email: foo@bar.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: asy.fhir.io
basePath: /0.0.1
schemes:
  - https
securityDefinitions:
  standard_auth:
    type: oauth2
    authorizationUrl: http://www.some.auth.url/com/authorize
    flow: implicit
    scopes:
      read:patients: Read some patient data
paths:
  /metadata:
    get:
      tags:
        - metadata
      summary: Get metadata for SMART on FHIR authorization access
      description: Returns the conformance statement and urls necessary for authentication
      produces:
        - application/json
        - application/xml
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Metadata'
        400:
          description: Invalid parameter
          schema:
            $ref: '#/definitions/Invalid'
        401:
          description: Unauthorized request
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: Insufficient scope
          schema:
            $ref: '#/definitions/InsufficientScope'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/NotFound'
        410:
          description: Resource deleted
          schema:
            $ref: '#/definitions/Deleted'
  /Patient/{id}:
    get:
      tags:
        - patient
      summary: Get patient by their patient id
      description: Returns a single patient
      security:
        standard_auth:
          - read:patients
      produces:
        - application/json
        - application/xml
      parameters:
        -
          name: id
          in: path
          description: Id of the patient to look up
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Patient'
        400:
          description: Invalid parameter
          schema:
            $ref: '#/definitions/Invalid'
        401:
          description: Unauthorized request
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: Insufficient scope
          schema:
            $ref: '#/definitions/InsufficientScope'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/NotFound'
        410:
          description: Resource deleted
          schema:
            $ref: '#/definitions/Deleted'
definitions:
  Metadata:
    $ref: definitions/metadata.yml#/Metadata
  Patient:
    $ref: definitions/patient.yml#/Patient
  Invalid:
    $ref: definitions/errors.yml#/Invalid
  Unauthorized:
    $ref: definitions/errors.yml#/Unauthorized
  InsufficientScope:
    $ref: definitions/errors.yml#/InsufficientScope
  NotFound:
    $ref: definitions/errors.yml#/NotFound
  Deleted:
    $ref: definitions/errors.yml#/Deleted